+++ pkg/apis/extensions/v1alpha1/helper/helper_dnsrecord.go
@@ -1,0 +1,30 @@
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package helper
+
+import (
+	"net"
+
+	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+)
+
+// GetDNSRecordType returns the appropriate DNS record type (A/AAAA or CNAME) for the given address.
+func GetDNSRecordType(address string) extensionsv1alpha1.DNSRecordType {
+	if ip := net.ParseIP(address); ip != nil {
+		if ip.To4() != nil {
+			return extensionsv1alpha1.DNSRecordTypeA
+		}
+		return extensionsv1alpha1.DNSRecordTypeAAAA
+	}
+	return extensionsv1alpha1.DNSRecordTypeCNAME
+}
+
+// GetDNSRecordTTL returns the value of the given ttl, or 120 if nil.
+func GetDNSRecordTTL(ttl *int64) int64 {
+	if ttl != nil {
+		return *ttl
+	}
+	return 120
+}
